requirements.txt
pandas
requests
numpy
scikit-learn
python-dotenv
matplotlib
geopandas